pyproject.toml
[project]
name = "planecli"
version = "0.1.0"
description = "CLI for Plane.so project management"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Patrick Alves", email = "cpatrickalves@gmail.com" },
]
keywords = ["plane", "cli", "project-management"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Bug Tracking",
]
dependencies = [
    "cyclopts>=3.0",
    "rich>=13.0",
    "rapidfuzz>=3.0",
    "plane-sdk",
]

[project.scripts]
planecli = "planecli.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/planecli"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
